refactor: reorganize component items and enhance documentation structure
