+++ constants/componentItems/apmDashboards.ts
@@ -1,0 +1,19 @@
+import type { ComponentItem } from './types'
+
+export const APM_DASHBOARDS_ITEMS: ComponentItem[] = [
+  {
+    name: 'APM Metrics',
+    href: '/docs/dashboards/dashboard-templates/apm-metrics',
+    clickName: 'APM Metrics Dashboard Template',
+  },
+  {
+    name: 'Database Calls Monitoring',
+    href: '/docs/dashboards/dashboard-templates/db-calls-monitoring',
+    clickName: 'Database Calls Monitoring Dashboard Template',
+  },
+  {
+    name: 'HTTP API Monitoring',
+    href: '/docs/dashboards/dashboard-templates/http-api-monitoring',
+    clickName: 'HTTP API Monitoring Dashboard Template',
+  },
+]
