+++ constants/componentItems/apmQuickStart.ts
@@ -1,0 +1,52 @@
+import type { ComponentItem } from './types'
+
+export const APM_QUICK_START_ITEMS: ComponentItem[] = [
+  {
+    name: 'Python',
+    href: '/docs/instrumentation/python',
+    clickName: 'Python Instrumentation Link',
+  },
+  {
+    name: 'Java',
+    href: '/docs/instrumentation/java/overview',
+    clickName: 'Java Instrumentation Link',
+  },
+  {
+    name: 'JavaScript',
+    href: '/docs/instrumentation/javascript/overview',
+    clickName: 'JavaScript Instrumentation Link',
+  },
+  {
+    name: 'Golang (Go)',
+    href: '/docs/instrumentation/opentelemetry-golang',
+    clickName: 'Golang Instrumentation Link',
+  },
+  { name: 'PHP', href: '/docs/instrumentation/php', clickName: 'PHP Instrumentation Link' },
+  {
+    name: '.NET',
+    href: '/docs/instrumentation/opentelemetry-dotnet',
+    clickName: '.NET Instrumentation Link',
+  },
+  {
+    name: 'Ruby',
+    href: '/docs/instrumentation/ruby-on-rails',
+    clickName: 'Ruby on Rails Instrumentation Link',
+  },
+  {
+    name: 'Elixir',
+    href: '/docs/instrumentation/elixir',
+    clickName: 'Elixir Instrumentation Link',
+  },
+  { name: 'Rust', href: '/docs/instrumentation/rust', clickName: 'Rust Instrumentation Link' },
+  {
+    name: 'C++',
+    href: '/docs/instrumentation/opentelemetry-cpp',
+    clickName: 'C++ Instrumentation Link',
+  },
+  { name: 'Swift', href: '/docs/instrumentation/swift', clickName: 'Swift Instrumentation Link' },
+  {
+    name: 'Deno',
+    href: '/docs/instrumentation/opentelemetry-deno',
+    clickName: 'Deno Instrumentation Link',
+  },
+]
