+++ constants/componentItems/dashboardTemplates.ts
@@ -1,0 +1,364 @@
+import type { ComponentItem } from './types'
+
+export const DASHBOARD_TEMPLATES_ITEMS: ComponentItem[] = [
+  {
+    name: 'Agno',
+    href: '/docs/dashboards/dashboard-templates/agno-dashboard',
+    clickName: 'Agno Dashboard Template',
+  },
+  {
+    name: 'Amazon Bedrock',
+    href: '/docs/dashboards/dashboard-templates/amazon-bedrock-dashboard',
+    clickName: 'Amazon Bedrock Dashboard Template',
+  },
+  {
+    name: 'Anthropic API',
+    href: '/docs/dashboards/dashboard-templates/anthropic-dashboard',
+    clickName: 'Anthropic API Dashboard Template',
+  },
+  {
+    name: 'Apache Web Server',
+    href: '/docs/dashboards/dashboard-templates/apache-web-server',
+    clickName: 'Apache Web Server Dashboard Template',
+  },
+  {
+    name: 'APM',
+    href: '/docs/dashboards/dashboard-templates/apm-dashboards',
+    clickName: 'APM Dashboard Template',
+  },
+  {
+    name: 'ArgoCD',
+    href: '/docs/dashboards/dashboard-templates/argocd-dashboard',
+    clickName: 'ArgoCD Dashboard Template',
+  },
+  {
+    name: 'Autogen',
+    href: '/docs/dashboards/dashboard-templates/autogen-dashboard',
+    clickName: 'Autogen Dashboard Template',
+  },
+  {
+    name: 'AWS ElastiCache Redis',
+    href: '/docs/dashboards/dashboard-templates/aws-elasticache-redis',
+    clickName: 'AWS ElastiCache Redis Dashboard Template',
+  },
+  {
+    name: 'AWS RDS',
+    href: 'https://github.com/SigNoz/dashboards/tree/main/aws-rds',
+    clickName: 'AWS RDS Dashboard Template',
+  },
+  {
+    name: 'AWS SQS Prometheus',
+    href: 'https://github.com/SigNoz/dashboards/tree/main/aws-sqs-prometheus',
+    clickName: 'AWS SQS Prometheus Dashboard Template',
+  },
+  {
+    name: 'Azure OpenAI API',
+    href: '/docs/dashboards/dashboard-templates/azure-openai-dashboard',
+    clickName: 'Azure OpenAI API Dashboard Template',
+  },
+  {
+    name: 'CI/CD',
+    href: '/docs/dashboards/dashboard-templates/cicd',
+    clickName: 'CI/CD Dashboard Template',
+  },
+  {
+    name: 'Claude Code',
+    href: '/docs/dashboards/dashboard-templates/claude-code-dashboard/',
+    clickName: 'Claude Code Dashboard Template',
+  },
+  {
+    name: 'Claude Agent SDK',
+    href: '/docs/dashboards/dashboard-templates/claude-agent-sdk-dashboard/',
+    clickName: 'Claude Agent SDK Dashboard Template',
+  },
+  {
+    name: 'ClickHouse',
+    href: '/docs/dashboards/dashboard-templates/clickhouse-monitoring',
+    clickName: 'ClickHouse Dashboard Template',
+  },
+  {
+    name: 'Codex (OpenAI)',
+    href: '/docs/dashboards/dashboard-templates/codex-dashboard',
+    clickName: 'Codex Dashboard Template',
+  },
+  {
+    name: 'Cost Meter',
+    href: '/docs/dashboards/dashboard-templates/cost-meter',
+    clickName: 'Cost Meter Dashboard Template',
+  },
+  {
+    name: 'CouchDB',
+    href: '/docs/dashboards/dashboard-templates/couchdb',
+    clickName: 'CouchDB Dashboard Template',
+  },
+  {
+    name: 'Crew AI',
+    href: '/docs/dashboards/dashboard-templates/crewai-dashboard',
+    clickName: 'Crew AI Dashboard Template',
+  },
+  {
+    name: 'DeepSeek API',
+    href: '/docs/dashboards/dashboard-templates/deepseek-dashboard',
+    clickName: 'DeepSeek API Dashboard Template',
+  },
+  {
+    name: 'Docker Container Metrics',
+    href: '/docs/dashboards/dashboard-templates/docker-container-metrics',
+    clickName: 'Docker Container Metrics Dashboard Template',
+  },
+  {
+    name: 'ECS Infra Metrics',
+    href: 'https://github.com/SigNoz/dashboards/tree/main/ecs-infra-metrics',
+    clickName: 'ECS Infra Metrics Dashboard Template',
+  },
+  {
+    name: 'Envoy Proxy',
+    href: '/docs/dashboards/dashboard-templates/envoy-dashboard',
+    clickName: 'Envoy Proxy Dashboard Template',
+  },
+  {
+    name: 'Flask Monitoring',
+    href: '/docs/dashboards/dashboard-templates/flask-monitoring',
+    clickName: 'Flask Monitoring Dashboard Template',
+  },
+  {
+    name: 'Frontend Monitoring',
+    href: '/docs/dashboards/dashboard-templates/frontend-monitoring',
+    clickName: 'Frontend Monitoring Dashboard Template',
+  },
+  {
+    name: 'GCP Compute Engine',
+    href: 'https://github.com/SigNoz/dashboards/tree/main/gcp/compute-engine',
+    clickName: 'GCP Compute Engine Dashboard Template',
+  },
+  {
+    name: 'Google ADK',
+    href: '/docs/dashboards/dashboard-templates/google-adk-dashboard',
+    clickName: 'Google ADK Dashboard Template',
+  },
+  {
+    name: 'Google Gemini',
+    href: 'https://github.com/SigNoz/dashboards/tree/main/google-gemini',
+    clickName: 'Google Gemini Dashboard Template',
+  },
+  {
+    name: 'Go Runtime',
+    href: '/docs/dashboards/dashboard-templates/go-runtime-metrics',
+    clickName: 'Go Runtime Dashboard Template',
+  },
+  {
+    name: 'Grok',
+    href: '/docs/dashboards/dashboard-templates/grok-dashboard',
+    clickName: 'xAi Grok Dashboard Template',
+  },
+  {
+    name: 'Groq',
+    href: '/docs/dashboards/dashboard-templates/groq-dashboard',
+    clickName: 'Groq Dashboard Template',
+  },
+  {
+    name: 'Hadoop',
+    href: 'https://github.com/SigNoz/dashboards/tree/main/hadoop',
+    clickName: 'Hadoop Dashboard Template',
+  },
+  {
+    name: 'HAProxy',
+    href: '/docs/dashboards/dashboard-templates/haproxy-monitoring',
+    clickName: 'HAProxy Dashboard Template',
+  },
+  {
+    name: 'Haystack',
+    href: '/docs/dashboards/dashboard-templates/haystack-dashboard',
+    clickName: 'Haystack Dashboard Template',
+  },
+  {
+    name: 'Host Metrics',
+    href: '/docs/dashboards/dashboard-templates/hostmetrics-dashboards',
+    clickName: 'Host Metrics Dashboard Template',
+  },
+  {
+    name: 'Hugging Face',
+    href: '/docs/dashboards/dashboard-templates/huggingface-dashboard',
+    clickName: 'Hugging Face Dashboard Template',
+  },
+  {
+    name: 'Inkeep',
+    href: '/docs/dashboards/dashboard-templates/inkeep-dashboard',
+    clickName: 'Inkeep Dashboard Template',
+  },
+  {
+    name: 'Jenkins',
+    href: 'https://github.com/SigNoz/dashboards/tree/main/jenkins',
+    clickName: 'Jenkins Dashboard Template',
+  },
+  {
+    name: 'JMX',
+    href: 'https://github.com/SigNoz/dashboards/tree/main/jmx',
+    clickName: 'JMX Dashboard Template',
+  },
+  {
+    name: 'JVM',
+    href: '/docs/dashboards/dashboard-templates/jvm-metrics',
+    clickName: 'JVM Dashboard Template',
+  },
+  {
+    name: 'Kubernetes',
+    href: '/docs/dashboards/dashboard-templates/kubernetes-dashboards',
+    clickName: 'K8s Infra Metrics Dashboard Template',
+  },
+  {
+    name: 'KEDA',
+    href: 'https://github.com/SigNoz/dashboards/tree/main/keda',
+    clickName: 'KEDA Dashboard Template',
+  },
+  {
+    name: 'Key Operations',
+    href: '/docs/dashboards/dashboard-templates/key-operations',
+    clickName: 'Key Operations Dashboard Template',
+  },
+  {
+    name: 'LiteLLM',
+    href: '/docs/dashboards/dashboard-templates/litellm-dashboards',
+    clickName: 'LiteLLM Dashboard Template',
+  },
+  {
+    name: 'LiveKit',
+    href: '/docs/dashboards/dashboard-templates/livekit-dashboard',
+    clickName: 'LiveKit Dashboard Template',
+  },
+  {
+    name: 'LLM Observability',
+    href: 'https://github.com/SigNoz/dashboards/tree/main/llm-observability',
+    clickName: 'LLM Observability Dashboard Template',
+  },
+  {
+    name: 'Mastra',
+    href: '/docs/dashboards/dashboard-templates/mastra-dashboard',
+    clickName: 'Mastra Dashboard Template',
+  },
+  {
+    name: 'Memcached',
+    href: '/docs/dashboards/dashboard-templates/memcached',
+    clickName: 'Memcached Dashboard Template',
+  },
+  {
+    name: 'Mistral AI',
+    href: '/docs/dashboards/dashboard-templates/mistral-dashboard',
+    clickName: 'Mistral AI Dashboard Template',
+  },
+  {
+    name: 'MongoDB',
+    href: 'https://github.com/SigNoz/dashboards/tree/main/mongodb',
+    clickName: 'MongoDB Dashboard Template',
+  },
+  {
+    name: 'MySQL',
+    href: '/docs/dashboards/dashboard-templates/mysql',
+    clickName: 'MySQL Dashboard Template',
+  },
+  {
+    name: 'Nginx',
+    href: '/docs/dashboards/dashboard-templates/nginx',
+    clickName: 'Nginx Dashboard Template',
+  },
+  {
+    name: 'NVIDIA DCGM',
+    href: 'https://github.com/SigNoz/dashboards/tree/main/nvidia-dcgm',
+    clickName: 'NVIDIA DCGM Dashboard Template',
+  },
+  {
+    name: 'Nomad',
+    href: 'https://github.com/SigNoz/dashboards/tree/main/nomad',
+    clickName: 'Nomad Dashboard Template',
+  },
+  {
+    name: 'Ollama',
+    href: '/docs/dashboards/dashboard-templates/ollama-dashboard',
+    clickName: 'Ollama Dashboard Template',
+  },
+  {
+    name: 'OpenAI',
+    href: '/docs/dashboards/dashboard-templates/openai-dashboard',
+    clickName: 'OpenAI Dashboard Template',
+  },
+  {
+    name: 'OpenClaw',
+    href: '/docs/dashboards/dashboard-templates/openclaw-dashboard',
+    clickName: 'OpenClaw Dashboard Template',
+  },
+  {
+    name: 'OpenRouter',
+    href: '/docs/dashboards/dashboard-templates/openrouter-dashboard',
+    clickName: 'OpenRouter Dashboard Template',
+  },
+  {
+    name: 'Pipecat',
+    href: '/docs/dashboards/dashboard-templates/pipecat-dashboard',
+    clickName: 'Pipecat Dashboard Template',
+  },
+  {
+    name: 'PostgreSQL',
+    href: '/docs/dashboards/dashboard-templates/postgresql',
+    clickName: 'PostgreSQL Dashboard Template',
+  },
+  {
+    name: 'Pydantic AI',
+    href: '/docs/dashboards/dashboard-templates/pydantic-ai-dashboard',
+    clickName: 'Pydantic AI Dashboard Template',
+  },
+  {
+    name: 'RabbitMQ',
+    href: '/docs/dashboards/dashboard-templates/rabbitmq',
+    clickName: 'RabbitMQ Dashboard Template',
+  },
+  {
+    name: 'Redis',
+    href: 'https://github.com/SigNoz/dashboards/tree/main/redis',
+    clickName: 'Redis Dashboard Template',
+  },
+  {
+    name: 'Render',
+    href: '/docs/dashboards/dashboard-templates/render-dashboard',
+    clickName: 'Render Dashboard Template',
+  },
+  {
+    name: 'Semantic Kernel',
+    href: '/docs/dashboards/dashboard-templates/semantic-kernel-dashboard',
+    clickName: 'Semantic Kernel Dashboard Template',
+  },
+  {
+    name: 'SigNoz Ingestion Analysis',
+    href: '/docs/dashboards/dashboard-templates/signoz-ingestion-analysis',
+    clickName: 'SigNoz Ingestion Analysis Dashboard Template',
+  },
+  {
+    name: 'SLURM',
+    href: 'https://github.com/SigNoz/dashboards/tree/main/slurm',
+    clickName: 'SLURM Dashboard Template',
+  },
+  {
+    name: 'Snowflake',
+    href: 'https://github.com/SigNoz/dashboards/tree/main/snowflake',
+    clickName: 'Snowflake Dashboard Template',
+  },
+  {
+    name: 'Supabase',
+    href: '/docs/dashboards/dashboard-templates/supabase',
+    clickName: 'Supabase Dashboard Template',
+  },
+  {
+    name: 'Temporal Agents',
+    href: '/docs/dashboards/dashboard-templates/temporal-dashboard',
+    clickName: 'Temporal Agents Dashboard Template',
+  },
+  {
+    name: 'Temporal.io',
+    href: 'https://github.com/SigNoz/dashboards/tree/main/temporal.io',
+    clickName: 'Temporal.io Dashboard Template',
+  },
+  {
+    name: 'Vercel AI SDK',
+    href: 'https://github.com/SigNoz/dashboards/tree/main/vercel-ai-sdk',
+    clickName: 'Vercel AI SDK Dashboard Template',
+  },
+]
