+++ constants/componentItems/hostMetricsDashboards.ts
@@ -1,0 +1,14 @@
+import type { ComponentItem } from './types'
+
+export const HOST_METRICS_DASHBOARDS_ITEMS: ComponentItem[] = [
+  {
+    name: 'Host Metrics K8s',
+    href: '/docs/dashboards/dashboard-templates/hostmetrics-k8s',
+    clickName: 'Host Metrics K8s Dashboard Template',
+  },
+  {
+    name: 'Host Metrics (VM)',
+    href: '/docs/dashboards/dashboard-templates/hostmetrics-vm',
+    clickName: 'Host Metrics VM Dashboard Template',
+  },
+]
