+++ constants/componentItems/hostingDecision.ts
@@ -1,0 +1,14 @@
+import type { ComponentItem } from './types'
+
+export const HOSTING_DECISION_ITEMS: ComponentItem[] = [
+  {
+    name: 'Compare Self Host vs Cloud',
+    href: '/blog/cloud-vs-self-hosted-deployment-guide/',
+    clickName: 'Compare Self Host vs Cloud',
+  },
+  {
+    name: 'Get Started - Free',
+    href: '/teams/',
+    clickName: 'Get Started - Free',
+  },
+]
