+++ constants/componentItems/kubernetesDashboards.ts
@@ -1,0 +1,44 @@
+import type { ComponentItem } from './types'
+
+export const KUBERNETES_DASHBOARDS_ITEMS: ComponentItem[] = [
+  {
+    name: 'Kubernetes Cluster Metrics (Overview)',
+    href: '/docs/dashboards/dashboard-templates/kubernetes-cluster-metrics',
+    clickName: 'Kubernetes Cluster Metrics Dashboard Template',
+  },
+  {
+    name: 'Kubernetes Events',
+    href: '/docs/dashboards/dashboard-templates/kubernetes-events',
+    clickName: 'Kubernetes Events Dashboard Template',
+  },
+  {
+    name: 'Kubernetes Pod Metrics (Overall)',
+    href: '/docs/dashboards/dashboard-templates/kubernetes-pod-metrics-overall',
+    clickName: 'Kubernetes Pod Metrics Overall Dashboard Template',
+  },
+  {
+    name: 'Kubernetes Pod Metrics (Detailed)',
+    href: '/docs/dashboards/dashboard-templates/kubernetes-pod-metrics-detailed',
+    clickName: 'Kubernetes Pod Metrics Detailed Dashboard Template',
+  },
+  {
+    name: 'Kubernetes Node Metrics (Overall)',
+    href: '/docs/dashboards/dashboard-templates/kubernetes-node-metrics-overall',
+    clickName: 'Kubernetes Node Metrics Overall Dashboard Template',
+  },
+  {
+    name: 'Kubernetes Node Metrics (Detailed)',
+    href: '/docs/dashboards/dashboard-templates/kubernetes-node-metrics-detailed',
+    clickName: 'Kubernetes Node Metrics Detailed Dashboard Template',
+  },
+  {
+    name: 'Kubernetes PVC Metrics',
+    href: '/docs/dashboards/dashboard-templates/kubernetes-pvc',
+    clickName: 'Kubernetes PVC Metrics Dashboard Template',
+  },
+  {
+    name: 'Host Metrics',
+    href: '/docs/dashboards/dashboard-templates/hostmetrics-k8s',
+    clickName: 'Host Metrics Dashboard Template',
+  },
+]
