+++ constants/componentItems/litellmDashboards.ts
@@ -1,0 +1,14 @@
+import type { ComponentItem } from './types'
+
+export const LITELLM_DASHBOARDS_ITEMS: ComponentItem[] = [
+  {
+    name: 'LiteLLM Proxy',
+    href: '/docs/dashboards/dashboard-templates/litellm-proxy-dashboard',
+    clickName: 'LiteLLM Proxy Dashboard Template',
+  },
+  {
+    name: 'LiteLLM SDK',
+    href: '/docs/dashboards/dashboard-templates/litellm-sdk-dashboard',
+    clickName: 'LiteLLM SDK Dashboard Template',
+  },
+]
