+++ constants/componentItems/llmMonitoring.ts
@@ -1,0 +1,107 @@
+import type { ComponentItem } from './types'
+
+export const LLM_MONITORING_ITEMS: ComponentItem[] = [
+  { name: 'Agno', href: '/docs/agno-monitoring', clickName: 'Agno Monitoring' },
+  {
+    name: 'Amazon Bedrock',
+    href: '/docs/amazon-bedrock-monitoring',
+    clickName: 'Amazon Bedrock Monitoring',
+  },
+  {
+    name: 'Anthropic API',
+    href: '/docs/anthropic-monitoring',
+    clickName: 'Anthropic API Monitoring',
+  },
+  { name: 'AutoGen', href: '/docs/autogen-observability', clickName: 'AutoGen Monitoring' },
+  {
+    name: 'Azure OpenAI API',
+    href: '/docs/azure-openai-monitoring',
+    clickName: 'Azure OpenAI API Monitoring',
+  },
+  {
+    name: 'Claude Code',
+    href: '/docs/claude-code-monitoring',
+    clickName: 'Claude Code Monitoring',
+  },
+  {
+    name: 'Claude Agent SDK',
+    href: '/docs/claude-agent-monitoring',
+    clickName: 'Claude Agent SDK Monitoring',
+  },
+  {
+    name: 'Codex (OpenAI)',
+    href: '/docs/codex-monitoring',
+    clickName: 'Codex (OpenAI) Monitoring',
+  },
+  { name: 'Crew AI', href: '/docs/crewai-observability', clickName: 'Crew AI Monitoring' },
+  { name: 'DeepSeek API', href: '/docs/deepseek-monitoring', clickName: 'DeepSeek Monitoring' },
+  {
+    name: 'Google ADK',
+    href: '/docs/google-adk-observability',
+    clickName: 'Google ADK Monitoring',
+  },
+  {
+    name: 'Google Gemini',
+    href: '/docs/google-gemini-monitoring',
+    clickName: 'Google Gemini Monitoring',
+  },
+  { name: 'Grok', href: '/docs/grok-monitoring', clickName: 'Grok Monitoring' },
+  { name: 'Groq', href: '/docs/groq-observability', clickName: 'Groq Monitoring' },
+  { name: 'Haystack', href: '/docs/haystack-monitoring', clickName: 'Haystack Monitoring' },
+  {
+    name: 'Hugging Face',
+    href: '/docs/huggingface-observability',
+    clickName: 'Hugging Face Monitoring',
+  },
+  { name: 'Inkeep', href: '/docs/inkeep-monitoring', clickName: 'Inkeep Monitoring' },
+  {
+    name: 'LangChain/LangGraph',
+    href: '/docs/langchain-observability',
+    clickName: 'LangChain Monitoring',
+  },
+  { name: 'Langtrace', href: '/docs/langtrace', clickName: 'Langtrace' },
+  { name: 'LiteLLM', href: '/docs/litellm-observability', clickName: 'LiteLLM Monitoring' },
+  { name: 'LiveKit', href: '/docs/livekit-observability', clickName: 'LiveKit Monitoring' },
+  {
+    name: 'LlamaIndex',
+    href: '/docs/llamaindex-observability',
+    clickName: 'LlamaIndex Monitoring',
+  },
+  { name: 'Mastra', href: '/docs/mastra-observability', clickName: 'Mastra Monitoring' },
+  {
+    name: 'Mistral AI',
+    href: '/docs/mistral-observability',
+    clickName: 'Mistral AI Monitoring',
+  },
+  { name: 'Ollama', href: '/docs/ollama-monitoring', clickName: 'Ollama Monitoring' },
+  { name: 'OpenAI', href: '/docs/openai-monitoring', clickName: 'OpenAI Monitoring' },
+  { name: 'OpenClaw', href: '/docs/openclaw-monitoring', clickName: 'OpenClaw Monitoring' },
+  { name: 'OpenLIT', href: '/docs/openlit', clickName: 'OpenLIT' },
+  {
+    name: 'OpenRouter',
+    href: '/docs/openrouter-observability',
+    clickName: 'OpenRouter Monitoring',
+  },
+  { name: 'Pipecat', href: '/docs/pipecat-monitoring', clickName: 'Pipecat Monitoring' },
+  {
+    name: 'Pydantic AI',
+    href: '/docs/pydantic-ai-observability',
+    clickName: 'Pydantic AI Monitoring',
+  },
+  {
+    name: 'Semantic Kernel',
+    href: '/docs/semantic-kernel-observability',
+    clickName: 'Semantic Kernel Monitoring',
+  },
+  { name: 'Temporal', href: '/docs/temporal-observability', clickName: 'Temporal Monitoring' },
+  {
+    name: 'Traceloop (OpenLLMetry)',
+    href: '/docs/traceloop',
+    clickName: 'Traceloop (OpenLLMetry)',
+  },
+  {
+    name: 'Vercel AI SDK',
+    href: '/docs/vercel-ai-sdk-observability',
+    clickName: 'Vercel AI SDK Monitoring',
+  },
+]
