+++ constants/componentItems/migrateToSignoz.ts
@@ -1,0 +1,39 @@
+import type { ComponentItem } from './types'
+
+export const MIGRATE_TO_SIGNOZ_ITEMS: ComponentItem[] = [
+  {
+    name: 'Migrate from Datadog',
+    href: '/docs/migration/migrate-from-datadog-to-signoz/',
+    clickName: 'Migrate from Datadog',
+  },
+  {
+    name: 'Migrate from Grafana',
+    href: '/docs/migration/migrate-from-grafana/',
+    clickName: 'Migrate from Grafana',
+  },
+  {
+    name: 'Migrate from ELK',
+    href: '/docs/migration/migrate-from-elk/',
+    clickName: 'Migrate from ELK',
+  },
+  {
+    name: 'Migrate from New Relic',
+    href: '/docs/migration/migrate-from-newrelic/',
+    clickName: 'Migrate from New Relic',
+  },
+  {
+    name: 'Migrate from Honeycomb',
+    href: '/docs/migration/migrate-from-honeycomb-to-signoz/',
+    clickName: 'Migrate from Honeycomb',
+  },
+  {
+    name: 'Migrate from OpenTelemetry',
+    href: '/docs/migration/migrate-from-opentelemetry-to-signoz/',
+    clickName: 'Migrate from OpenTelemetry',
+  },
+  {
+    name: 'Migrate from Self-Hosted SigNoz',
+    href: '/docs/migration/migrate-to-signoz-cloud/',
+    clickName: 'Migrate from Self-Hosted SigNoz',
+  },
+]
