+++ constants/componentItems/types.ts
@@ -1,0 +1,5 @@
+export interface ComponentItem {
+  name: string
+  href: string
+  clickName: string
+}
