+++ constants/componentItems/webVitals.ts
@@ -1,0 +1,14 @@
+import type { ComponentItem } from './types'
+
+export const WEB_VITALS_ITEMS: ComponentItem[] = [
+  {
+    name: 'Sending Web Vitals as Metrics',
+    href: '/docs/frontend-monitoring/web-vitals-with-metrics',
+    clickName: 'Sending Web Vitals as Metrics',
+  },
+  {
+    name: 'Sending Web Vitals as Traces',
+    href: '/docs/frontend-monitoring/web-vitals-with-traces',
+    clickName: 'Sending Web Vitals as Traces',
+  },
+]
